refactor(frontend): move graphql queries to separate file
