+++ React_Project/vite.config.ts
@@ -1,10 +1,10 @@
-import { sentryVitePlugin } from "@sentry/vite-plugin";
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
 import tailwindcss from "tailwindcss";
+
+import react from "@vitejs/plugin-react";
+import { sentryVitePlugin } from "@sentry/vite-plugin";
 import { VitePWA } from "vite-plugin-pwa";
 
-// https://vitejs.dev/config/
 export default defineConfig({
     plugins: [
         react(),
